+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * In S3 storage automatically write original filename to `Content-Disposition` header (janko-m)
 
+* Override `#to_s` in `Shrine::Attachment` for better introspection with `puts` (janko-m)
+
 ## 2.3.1 (2016-09-01)
 
 * Don't change permissions of existing directories in FileSystem storage (janko-m)